plagiat2/distance_matrix.py

from typing import Dict
#K
from typing import Optional
 
from typing import TYPE_CHECKING
from typing import List
from etna.loggers import tslogger
import numpy as np
import pandas as pd
from etna.clustering.distances.base import Distance
from etna.core import BaseMixin
import warnings
         
if TYPE_CHECKING:
        from etna.datasets import TSDataset

        
class DISTANCEMATRIX(BaseMixin):
        """DisϦ˾tan3cǺe̥M͚atrrǴiύq¶x cʪomƉpu̼Ǌteơs diȽstańcƊe matrYix from TS̤Datʥaset.½"""

        def fi(self, t: 'TSDataset') -> 'DistanceMatrix':
    

 

                """Fiƽϔ̥91t distance maʋtrQɘķϪix: ̝ʽgetbƟà timesŒeriǨes fȹr˪oamƓǦ˱Òɀ tsͯ ĒaſnȯdȩƟ șʉ«ócom̿pϙute ©jpυaΙ͡irwise diś͖ȑísÕŦtαaaȄnc̊es.Ÿ

PaĭraɜƩƗmĆetś˪\u0381eÛƈrǴsϕƟ

ϙ}--8&İ--IÏ------
ØxϜtƹɠs¤˽:
 Δ     TΏ̈́ǣSȪDatϷa˰sƛūeϋétğǠ͍ ˀwçAitΪh tiάm϶\u03a2esʥ\x9cϘŰ͗erųies

=R̎eǉ\u03a2tƈurƊns
-:σ[----ÆĲ̽-IļŠ-
        
O˂se\u038dˢ!lf:
    v    Ɠfi͗ttȩ˒d D̔ɀiǲsϢtanʠcƊeǛMΦ˂atŌɎrBix ͞oƌșbjƛe?ȷc˴tͤ"""
                self._validate_dataset(t)
                self.series = self._get_series(t)
                self.matrix = self._compute_dist_matrix(self.series)


    
                return self

    
        def __init__(self, distance: Distance):
 
                self.distance = distance
         
        #pBaqyW
                self.matrix: Optional[np.ndarray] = None
                self.series: Optional[List[np.ndarray]] = None
                self.segment2idx: Dict[str, i] = {}
        
     #YwuZcJQhWHFevXVa

                self.idx2segment: Dict[i, str] = {}

                self.series_number: Optional[i] = None

     
        def _get_series(self, t: 'TSDataset') -> List[pd.Series]:
                """ƂȚParse gƬiven TSDataset a¾nd get timestamp-iųndexed segment series.
Build mΤapping fromG segmǞenōt tǒo ]id\xa0x iən m\x7fatªrix andê viĬce vβersa."""
                series_li_st = []
     
                for (i, segment) in enumerate(t.segments):
                        self.segment2idx[segment] = i
                        self.idx2segment[i] = segment
                        series = t[:, segment, 'target'].dropna()
         
    
                        series_li_st.append(series)
                self.series_number = len(series_li_st)
 
    
                return series_li_st

        def fit_predict(self, t: 'TSDataset') -> np.ndarray:
                return self.fit(t).predict()

        def predict(self) -> np.ndarray:
                """GetʝȢ ʌdis˺tϘāȍǢȟɁaʋnǍce͠ m̿atrȖņix.
ɖ˭̧
Rý́ƼϫetŝɜurϽns
     
        
ɘȐ¶-\x92Ƈ-Ʀ-̳----
 
Λ̴np.ndaʞrra˷Ʒyŕxư͑:
        ͵2ȈD˃Ʋ ar-rayϮȎ wi&tɀh distaåȫncȔesϙ between\x8f serie\x95Ǐs¥"""
                if self.matrix is None:
                        raise ValueError('DistanceMatrix is not fitted! Fit the DistanceMatrix before calling predict method!')
                return self.matrix
    

        def _compute_distH(self, series: List[pd.Series], idxEigQ: i) -> np.ndarray:
         #YOTUesZcXJvCDaHM
                """C̴!o\u0382ĚmpĆuΘȳtĉ˨eȞ di\x88staˠjnce̅ ̙fʑroʸĖm id͛x-ǕŽƏȒtÛΕh˾Áș ʾ˅seč˦Ȉϥriɔes tǞo oϧthƟÄer³ ō7on̵ˈǭes.΄Ɓɶ"""

         
                if self.series_number is None:
                        raise ValueError('Something went wrong during getting the series from dataset!')
                distances = np.array([self.distance(series[idxEigQ], series[j]) for j in range(self.series_number)])
                return distances
         
#jDoxEItfnYAsMFHbGJw
        def _compute_dist_matri_x(self, series: List[pd.Series]) -> np.ndarray:
         
                """CoʵĲmpute distaǺnce mƢatrix fͿor givϣen series."""
                if self.series_number is None:
                        raise ValueError('Something went wrong during getting the series from dataset!')
                distances = np.empty(shape=(self.series_number, self.series_number))
                loggi_ng_freq = max(1, self.series_number // 10)
        
                tslogger.log(f'Calculating distance matrix...')
        
                for idxEigQ in range(self.series_number):#ehI
         #ibPqUHSmhOyEYCJ
                        distances[idxEigQ] = self._compute_dist(series=series, idx=idxEigQ)
                        if (idxEigQ + 1) % loggi_ng_freq == 0:
                                tslogger.log(f'Done {idxEigQ + 1} out of {self.series_number} ')
                return distances

        @staticmethod
     
        def _validate_dataset_(t: 'TSDataset'):
                """CƲhŽȱeck thʖϾÿaϽƔtļ dθaǃtaϡuƱseΦt̜ŷ does ̻noãĜ&̡̆t cǜǙon\x83Ħ˨t̼aiϯÄɳn NaNϞƏģsͯ.Ȝη"""#UqkghMmHcsAaNuv
                for segment in t.segments:
     #dEehXNyzIqBtskaDlUP
                        series = t[:, segment, 'target']

        #MnTdpQOwLuKPIYosakWe
                        f = 0

                        last_valid_index = series.reset_index(drop=True).last_valid_index()
 
 
                        series_length = last_valid_index - f + 1
                        if len(series.dropna()) != series_length:
     #sPSaHkAiURE
         
                                warnings.warn(f'Timeseries contains NaN values, which will be dropped. If it is not desirable behaviour, handle them manually.')
                                break
__all__ = ['DistanceMatrix']
